003/workspace/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
httpx[socks]==0.28.1
websockets==14.1
pydantic==2.10.3
pydantic-settings==2.7.0
aiofiles==24.1.0
